Little documentation, API and separation of file
